+++ Outback/src/main/java/com/human/outback/DTO/Menu.java
@@ -1,4 +1,4 @@
-package com.human.outback;
+package com.human.outback.DTO;
 
 public class Menu {
 	private int menu_code;
